+++ src/main/scala/org/mmisw/orr/client/ont/jenaHelper.scala
@@ -8,6 +8,7 @@ import org.apache.jena.rdf.model.ModelFactory
 object jenaHelper {
 
   def loadOntModel(uri: String, fileOpt: Option[File] = None): OntModel = {
+    //println(s"loadOntModel: uri=$uri fileOpt=$fileOpt")
     val spec = new OntModelSpec(OntModelSpec.OWL_MEM)
     spec.setDocumentManager(new OntDocumentManager)
     val model = ModelFactory.createOntologyModel(spec, null)
@@ -15,9 +16,11 @@ object jenaHelper {
     model.getDocumentManager.setProcessImports(false)
     fileOpt match {
       case None       ⇒ model.read(uri);
+
       case Some(file) ⇒
+      val lang = if (file.getName.endsWith(".ttl")) "TTL" else "RDF/XML"
         val is = new FileInputStream(file)
-        try model.read(is, uri)
+        try model.read(is, uri, lang)
         finally is.close()
     }
     model